feat(session): Auto-login data separation

+++ src/main/java/mingzuozhibi/action/SessionController.java
@@ -22,6 +22,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.servlet.http.HttpSession;
 import java.time.LocalDateTime;
 import java.util.Set;
 import java.util.UUID;
@@ -61,14 +62,14 @@ public class SessionController extends BaseController {
         String header = getAttributes().getRequest().getHeader("X-AUTO-LOGIN");
         if (header == null || header.isEmpty()) {
             if (LOGGER.isDebugEnabled()) {
-                debugRequest("[用户自动登入: 未发现有效的X-AUTO-LOGIN]");
+                debugRequest("[自动登入: 未发现有效的X-AUTO-LOGIN]");
             }
             return false;
         }
 
         if (header.length() != 36) {
             if (LOGGER.isWarnEnabled()) {
-                warnRequest("[用户自动登入: 异常的X-AUTO-LOGIN][token={}]", header);
+                warnRequest("[自动登入: 异常的X-AUTO-LOGIN][token={}]", header);
             }
             return false;
         }
@@ -76,7 +77,7 @@ public class SessionController extends BaseController {
         AutoLogin autoLogin = dao.lookup(AutoLogin.class, "token", header);
         if (autoLogin == null) {
             if (LOGGER.isDebugEnabled()) {
-                debugRequest("[用户自动登入: 服务器未找到相应数据][token={}]", header);
+                debugRequest("[自动登入: 服务器未找到相应数据][token={}]", header);
             }
             return false;
         }
@@ -84,7 +85,7 @@ public class SessionController extends BaseController {
         String username = autoLogin.getUser().getUsername();
         if (autoLogin.getExpired().isBefore(LocalDateTime.now())) {
             if (LOGGER.isDebugEnabled()) {
-                debugRequest("[用户自动登入: TOKEN已过期][username={}][expired={}]",
+                debugRequest("[自动登入: TOKEN已过期][username={}][expired={}]",
                         username, autoLogin.getExpired());
             }
             return false;
@@ -92,7 +93,7 @@ public class SessionController extends BaseController {
 
         if (!autoLogin.getUser().isEnabled()) {
             if (LOGGER.isInfoEnabled()) {
-                infoRequest("[用户自动登入: 用户已被停用][username={}]", username);
+                infoRequest("[自动登入: 用户已被停用][username={}]", username);
             }
             return false;
         }
@@ -101,8 +102,12 @@ public class SessionController extends BaseController {
         doLoginSuccess(userDetails.getPassword(), userDetails);
         onLoginSuccess(username, false);
 
+        HttpSession httpSession = getAttributes().getRequest().getSession();
+        httpSession.setAttribute("autoLoginId", autoLogin.getId());
+
         if (LOGGER.isInfoEnabled()) {
-            infoRequest("[用户自动登入: 已成功自动登入][username={}]", username);
+            infoRequest("[自动登入: 已成功自动登入][username={}][autoLoginId={}]",
+                    username, autoLogin.getId());
         }
 
         return true;
@@ -166,7 +171,16 @@ public class SessionController extends BaseController {
                 getAttributes().getResponse().addHeader("X-AUTO-LOGIN", header);
 
                 LocalDateTime expired = LocalDateTime.now().withNano(0).plusDays(14);
-                dao.save(new AutoLogin(user, header, expired));
+                AutoLogin autoLogin = new AutoLogin(user, header, expired);
+
+                dao.save(autoLogin);
+
+                HttpSession httpSession = getAttributes().getRequest().getSession();
+                httpSession.setAttribute("autoLoginId", autoLogin.getId());
+
+                if (LOGGER.isDebugEnabled()) {
+                    LOGGER.debug("[已设置自动登入][autoLoginId={}]", autoLogin.getId());
+                }
             }
         });
     }
@@ -180,15 +194,18 @@ public class SessionController extends BaseController {
                 UUID.randomUUID().toString(), "Guest", GUEST_AUTHORITIES)
         );
 
-        dao.execute(session -> {
-            User user = dao.lookup(User.class, "username", authentication.getName());
-            dao.findBy(AutoLogin.class, "user", user).forEach(autoLogin -> {
+        HttpSession httpSession = getAttributes().getRequest().getSession();
+        Long autoLoginId = (Long) httpSession.getAttribute("autoLoginId");
+
+        if (autoLoginId != null) {
+            dao.execute(session -> {
+                AutoLogin autoLogin = dao.get(AutoLogin.class, autoLoginId);
                 if (LOGGER.isDebugEnabled()) {
-                    debugRequest("[用户登出: 删除自动登入数据][token={}]", autoLogin.getToken());
+                    debugRequest("[用户登出: 删除自动登入数据][autoLoginId={}]", autoLogin.getId());
                 }
                 dao.delete(autoLogin);
             });
-        });
+        }
 
         getAttributes().getResponse().addHeader("X-AUTO-LOGIN", "");
